backend/requirements.txt
# Horizontal Decision Platform - Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
greenlet==3.0.3

# Caching & Queue
redis==5.0.1
celery==5.3.6

# Object Storage
boto3==1.34.34
minio==7.2.3

# AI/ML (optional - use mock by default)
openai==1.10.0
anthropic==0.18.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.1.2

# PDF & Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
httpx==0.26.0
aiofiles==23.2.1

# Monitoring & Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.5.1

# Development
black==23.12.1
isort==5.13.2
mypy==1.8.0
